backend/requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
frozendict==2.4.6
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multitasking==0.0.11
numpy==2.2.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
yfinance==0.2.52
